Added rendering, color struct, and renamed Logging
