+++ color.h
@@ -1,0 +1,9 @@
+#include <SDL2/SDL.h>
+struct color {
+  Uint8 r;
+  Uint8 g;
+  Uint8 b;
+  Uint8 a;
+};
+color BLACK = {0, 0, 0, 255};
+color WHITE = {255, 255, 255, 255};
